requirements.txt
accelerate==1.6.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
chronos-forecasting @ git+https://github.com/amazon-science/chronos-forecasting.git@6a9c8dadac04eb85befc935043e3e2cce914267f
colorama==0.4.6
colorlog==6.9.0
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nixtlats==0.5.2
numpy==2.2.5
openpyxl==3.1.5
optuna==4.3.0
packaging==25.0
pandas==2.2.3
properscoring==0.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
utilsforecast==0.2.12
